// src/models/profileInfo.model.ts
export interface ProfileInfo {
  name: string;
  ic: string;
  studentId: string;
  address: string;
  postcode: string;
  state: string;
  city: string;
  country: string;
  phone: string;
  gender: string;
  race: string;
  religion: string;
  email: string;
  course: string;
  courseName: string;
  dateRegistered: string;
  session: string;
  dateSemRegistered: string;
  sem: number;
  active: string;
  courseStatus: string;
  advisorName: string;
  dateBirth: string;
  placeBirth: string;
  status: string;
  citizen: string;
  nationality: string;
  kptmEmail: string;
  program: string;
}